Add part-way compatibility with Fixtures
Uses `FixtureSet.identify` to have the same logic as the primary key generation that fixtures use, in that way
we're a drop in replacement that could potentially be used alongside fixtures.

Since Fixtures issues a DELETE FROM for every table that's loaded with it, I'm opting to not support loading with Oaken and Fixtures on the same table.

E.g. using both `test/seeds.rb` (or `test/seeds/users.rb`) with `users.update` and `test/fixtures/users.yml` aren't supported.
Otherwise we'd need to override `setup_fixtures` which will have started transactions on the database connection already, so any data we create is rolled back after the test,
and we'd need probably complex connection management to handle that.