Rename comments to plans to help indicate potential data tables
Comments are usually attached to posts and their user creator's, whereas plans are
tables we want to ensure are in the database from the get-go.

I'm also swapping our test check to `after_save` to prevent a silent failure where the record
is already in the DB and then won't trigger `after_create`.