Make Oaken somewhat self-corrective
If we run the drop, create, migrate below the database has changed while Oaken still believes
the data is in the db and it's safe to replay the file when it's not.

The `find` will raise in that case and we can try to reset the file cache, so users can just rerun the command.

Test with:

```sh
RAILS_ENV=test bin/rails db:drop db:create db:migrate
bin/rails test
```

TODO: Hook into db:drop and reset our replay cache.