Allow loading SQL cases
In certain Rails versions there's been Marshal serialization bugs and it's pretty
tricky to test those because the new Rails version can't output the old Marshal structure.

So it can take a long time to reproduce, but now you can grab a dump from production and put it in a SQL file directly.

This also moves the cases folder into a test/ environment folder since that's what I'm recommending in the README.