Fix class level loading
When doing a `self.class.seed` from within a test, we'd be within the Test Class scope, so the `@loader` variable would assign there and
and not on `Oaken::Seeds.loader`.

The old structure was a little muddy anyway, so this attempts a rewrite that's hopefully clearer.

Now we'll always be within `Oaken::Seeds`' `singleton_class`.

Note: originally the point about injecting `load_onto Oaken::Seeds` was so you could potentially have other seeds modules, so you could have your app seeds split up ala Packwerk.
But it doesn't really work and it's kinda unclear how I'd make it work, so it's probably better to yank it in a bit.