Raise when `seed` finds no files. (#99)
* Raise when `seed` finds no files.

Ref https://github.com/kaspth/oaken/issues/95

In case users pass a typo'ed name/path to `seed` we won't give them any indication of what's going on.

Running a test fails silently.

Now we raise when we can't find any files.

* Split glob per path pattern

In case a user does `seed :accounts, :missing` we want to raise when we can't load from `:missing`.

Previously, we'd proceed silently because `:accounts` found files.