Don't use `table_name` + support `as:` on `register` (#111)
Oaken's seed files are meant to look like data scripts, which means the `users` in `users.create`,
uses a logical name that resolves to a User.

However, we'd come back around and try to use the `table_name` to derive the method name, which may have a prefix and/or suffix that we don't care about.
The developer using Oaken has already told us what name they expect because of how they triggered the `method_missing`,
so as long as we can resolve that into a type we're good.

If there's cases where that doesn't match, we'll also allow `register User, as: :aliased_users`
so you can have it however you want.