Remove explicit `seed :setup` mentions
Originally I put `:setup` as something users would have to call, but then I changed Oaken to automatically off an internal `seed :setup` once.

So the README was out-of-date and I forgot to remove the mentions.

Thanks to @ixti for catching it and mentioning it https://github.com/kaspth/oaken/issues/123#issuecomment-3349958943